fix error when deleting empty directories
With --ignore-fail-on-non-empty, non-emptiness is the only failure that
gets ignored by rmdir.
In the case that rmdir reaches $HOME and considers deleting it, it will
detect insufficient permissions and subsequently exit with an error,
even if $HOME is not empty.

Prevent this by calling rmdir with a relative path that excludes $HOME.