files: collect file definitions in internal list
1.
This fixes a bug where file definitions with the same attrname
but different file types collide:

  home.file.a = ...;
  xdg.configFile.a = ...;

The above config results in only one file being created, because when
the xdg.configFile attrset is internally assigned to home.file,
definitions with the same name are merged.

2.
It's a requirement for commit 'disallow setting source and text
simultaneously', because assigning xdg.configFile to home.file would
result in a double definition of "source" in case of "text == null".