qt: add option `qt.systemdServicePath`
This option contains a string suitable as value for the PATH
environment variable in systemd services that run Qt 5 applications.

Qt applications need this special treatment because they use PATH to
locate plugins that sometimes are necessary for proper functioning.

For now the option is not visible in the manual since the default
value is expected to work in most, if not all, cases.