bspwm: splitted the module and worked in the rules
Splitted the module in two different files: the `default.nix`,
which contains the function that writes the configuration to the
file. The other file, `options.nix`, contains all the module options.
The rules option is still WIP; it contains a copule of errors that
need to be fixed.

Huge thanks to https://github.com/chisui. He has been helping me a lot
with the errors the module has.