neovim: update types of arguments passed to wrapper
The neovim wrapper seems to have been updated, and
this broke the module's interface to the wrapper.
To reproduce the error simply enable neovim in HM
with an updated unstable channel.

In the case of extraPythonPackages & extraPython3Packages
the wrapper now expects functions that return
a list of packages; in order to preserve existing HM
configurations I made the mapping to a function in the
override call rather than changing the option type.

In the case of configure it seems a null value is no
longer accepted; I therefore updated the default to
the empty set {}.