Tmux: add module
This commit adds the tmux program to home-manager and wraps Tmux with
any `extraConfig` included in nix.

In addition to configuring Tmux, a user may specify tmux plugins from
nixpkgs. These can be included in the list of `plugins` and can either
be a package (all tmux plugins live under `nixpkgs.tmuxPlugins.*`), or
an object which includes the plugin and an `extraConfig` which will be
run immediately after sourcing the tmux plugin.

Finally, this commit introduces two nested programs which may be enabled
which depend on tmux: tmuxp and tmuxinator. These do not have the
ability to be configured, although this may be a future contribution.