home-manager: Add generation expiration command.
This adds a new command to home-manager's shell script that allows
generations to be removed that are older than an absolute or relative
date given.

This allows users to manage the expiration of their home-manager
generations separately from their system or user profiles via
nix-collect-garbage.  It's most important if user's desire to have
a convenient way to have different expiration times for home-manager
generations than other system or user profiles.