Exposes optionsJSON
This useful JSON list was previously inaccessible as a package or a module; at least as far as I could tell.
I suspect it was being used internally. It is incredibly useful for building auto-completion lists in various IDEs, as a navigation tool (combined with `jq`), and surely for other scripting tasks. I personally am using it to build a `home-manager-options` mimicking the existing `nixos-options` utility.