screen-locker: don't add trailing backslashes introduced by xautolockExtraOptions
systemd's unit file parser seems to have gotten a bit stricter, and with
systemd 242, the trailing backslash caused the next non-empty line to be
ignored.

In that case, this was [Section], so all subsequent settings were
mistakenly added to [Service], causing them to be ignored entirely.

Simplify and fix this by using concatStringsSep to build a single
ExecStart line.