mpd: use systemd journal instead of syslog
MPD is using syslog for its logging output, while it could directly log to
systemd's journal, as this daemon is primarily used as a systemd user service.
This change makes MPD log to standard output, which is captured by systemd.

See NixOS PR #57608, which does the same thing to NixOS's MPD service.