Experiment with flakes
I was curious to see what would be required to make home-manager work with
flakes, so I took a stab at it. I'm posting this here for those who are curious
about it.

It works surprisingly well. It delegate all the issues of finding nixpkgs and the
home-manager source tree to nix itself, so a lot of the `update-home` script can
be simplified.

To use it, create a new flake that provide a `homeConfiguration`, built using
`home-manager.lib.mkHome`, like so:

```
{
  edition = 201909;

  inputs.home-manager.uri = "/Users/georgesdubus/dev/perso/home-manager";

  outputs = { self, nixpkgs, home-manager }: {

    homeConfiguration = home-manager.lib.mkHome {
      imports = [ ./home.nix ];
      home.homeDirectory = "/Users/georgesdubus";
      home.username = "georgesdubus";
    };
  };
}

```
Then build the activation package with
`nix build .#homeConfiguration.activationPackage`. Or run the home-manager
script like so:
`nix run /Users/georgesdubus/dev/perso/home-manager -c home-manager build -f .`.
This assumes that you have a flake-enabled nix that you can get with `nix run nixpkgs.nixFlakes`.

Outstanding issues include:
- I've hardcoded it to darwin because that's the plaform I develop on, but it
  should be too much of a problem to fix that.
- Because the evaluation is pure, there is no way to access the $USER and $HOME
  environment variables, which means they need to be passed manually. I'm
  guessing it could be made nicer by having the wrapping script get them and
  pass them.
- Since the flake takes care of providing nixpkgs, it's not possible to provide
  configuration for it in `options.nixpkgs`. I believe this is still an open
  question for the flake RFC, so wait and see.
- Channels are not available, and nixos modules are not yet exposed from the
  nixpkgs flake, so the source of nixpkgs is passed manually to find the
  `assertions.nix` and `meta.nix` modules.
- I'm guessing we'll want to get closer to the behaviour of nixos-rebuild, once
  it's been finalized (multiple attributes in the `homeConfiguration` output,
  that kind of things)
- The manual didn't work out the box. I disabled it rather than looking into it.