files: Additional support for symlinked /nix
In the case where /nix is a link, for example, on macOS Catalina,
builtins.storeDir returns /nix, not the canonical location.

This causes tests on existing files to result in home-manager
thinking those files are outside of the store.

This change uses readlink on the store path so that the tests
work as intended.