home-manager: handle args with spaces to doBuildAttr
Presently, if you pass an argument with spaces in it to doBuildAttr,
it will be split it into multiple arguments to `nix build` or
`nix-build`. I came across this situation when I tried to use
home-manager on a system with spaces in `XDG_DATA_HOME`. Specifically,
the `home-manager` script errored out in trying to address the
read-news state file. With this change, I believe argument separation
should be preserved properly in `doBuildAttr`, and in fact the problem
seems fixed on my machine with this change.