tmux: use stable plugin names (name -> pname)
- supersedes f6afd95ef8dd587ca15c7090eb12d3cd3d08b08a
- fixes failed tests

In NixOS/nixpkgs@d5ccc59056b1ec8b012668191cefafcb2ec2653e, tmux plugins
were versioned and the `pname` attr was introduced in the derivations.

`pname` is the plain name and `name = "${pname}-${version}"`.

The tests failed in the time between the two commits because the plugin
names in `tmux.conf` were now unstable. The commit fixed this.