qutebrowser: add option to change the package used
This will allow users of Nix on non-NixOS to specify None as their
qutebrowser package, thereby allowing them to use home-manager to handle
their qutebrowser config, while running their system-wide installed copy
of qutebrowser instead of the Nix-installed one.

This is necessary, because of what is described in #1315: sometimes
Nix-installed Qutebrowser fails to start, because of GLX issues.

Fixes #1315.