kakoune: add support for plugins
The kakoune editor has a plugin mechanism and several plugins are
already packaged under `pkgs.kakounePlugins`. However, adding these
packages to `home.packages` is not enough: the `kakoune` package
needs to be configured with the list of plugins to include, so
that they get sourced on start-up.

We add a `programs.kakoune.plugins` option, analogous to
`programs.vim.plugins`.

The change is backwards compatible since `pkgs.kakoune` is defined as

```
wrapKakoune kakoune-unwrapped { };
```

and `wrapKakoune` defaults the list of plugins to empty.