systemd: use sd-switch
This makes the systemd module use the sd-switch application to perform
the unit switch during a generation activation.

Since the closure of sd-switch is relatively lightweight we
unconditionally pull it in as a dependency. We simultaneously remove
the `systemd.user.startServices` option and perform the switch action
automatically.

This is the second attempt at introducing this change, the first being
9c0fe3957bb91128d5d3599e289fe9344a293ece, which was quickly reverted
223e3c38a13fb45726c7a9d97e2612ae53ab4f98 since the feature didn't work
on systems where the user is not given a functional dbus session.
Since the user session is now always socket activated, see [0], we can
(hopefully) safely reintroduce the feature.

[0] https://github.com/NixOS/nixpkgs/pull/97801