lib.gvariant: mkRawString support
in dconf some things are easily represented by a string but must not
be a quoted string for example value `@mb true` describes a type maybe boolean and its value is true. currently home-manager is rendering this as '@mb true' which is stored in dconf but does not operate properly. using mkRawString "@mb true" does work as it writes a quoteless string in the ini