waybar: make settings a submodule
This changes the type of programs.waybar.settings from a list of submodules to
a submodule.

As far as I can tell, the only reason this was a list to begin with is that the
underlying/output JSON expects the top-level element to be a list, but I don't
see why we should let this limitation leak into the Nix code. By making this a
submodule instead of a list of submodules we make it much easier for users to
selectively override portions of their settings.